Added exit function to display_program Brightness now gets stored

+++ src/PIOMatrixOutput/color.cpp
@@ -6,7 +6,7 @@
 
 
 
-uint8_t globalBrightness = 4;
+uint32_t globalBrightness = 4;
 
 
 Color colorBlank(0,0,0);
